Simplified how ignored tokens are reported
Previously fetchNext() accepted a slice intended for ignored tokens. If the
slice was an nil, fetchNext() immediatelly called reportIgnoredtoken() on
each ignored token to avoid allocations. Otherwise, it appended tokens to
the slice nad returned it.

This change simplifies the implementation of fetchNext() and makes it always
appends ignored tokens.