Simplified how ignored tokens are reported (#49)
* Simplified how ignored tokens are reported

Previously fetchNext() accepted a slice intended for ignored tokens. If the
slice was an nil, fetchNext() immediatelly called reportIgnoredtoken() on
each ignored token to avoid allocations. Otherwise, it appended tokens to
the slice nad returned it.

This change simplifies the implementation of fetchNext() and makes it always
appends ignored tokens.

* Addressed review comments

The biggest change is in when pending tokens are reported. Specifically,
we only flush pending tokens

* before each call to fetchNext() unless p.next.symbol == noToken
* before assigning noToken to p.next.symbol
* at EOI
* before reporting a syntax error

This also ensures that no tokens are pending by the time parse() returns.

* Tabified lines using spaces in go_parser.ltp

* Moved the pending invariant to the Parser struct